SYNR-358

+++ R/exclusiveFileAccess.R
@@ -34,7 +34,10 @@ lockFile<-function(filePath, maxWaitSeconds=10) {
 
 unlockFile<-function(filePath) {
   lockdirPath <- lockdirPath(filePath)
-  unlink(lockdirPath)
+  file.remove(lockdirPath)
+  # the above doesn't work on cygwin, but one of these should
+  if (file.exists(lockdirPath)) unlink(lockdirPath)
+  if (file.exists(lockdirPath)) system(sprintf("rm -r %s", lockdirPath))
 }
 
 # return content of file
